+++ src/tests/client/client_test.cpp
@@ -96,7 +96,7 @@ class ConnectionImpl : public ConnectionBase
       m_state = StateDisconnected;
       m_pos = 0;
     }
-    virtual void getStatistics( int& /*totalIn*/, int& /*totalOut*/ ) {}
+    virtual void getStatistics( long int& /*totalIn*/, long int& /*totalOut*/ ) {}
     virtual ConnectionBase* newInstance() const { return 0; }
 
   private:
